+++ externals/kyuubi-spark-sql-engine/src/main/scala/org/apache/kyuubi/engine/spark/operation/ExecuteScala.scala
@@ -31,6 +31,7 @@ import org.apache.spark.sql.types.StructType
 import org.apache.kyuubi.KyuubiSQLException
 import org.apache.kyuubi.engine.spark.KyuubiSparkUtil._
 import org.apache.kyuubi.engine.spark.repl.KyuubiSparkILoop
+import org.apache.kyuubi.engine.spark.util.JsonUtils
 import org.apache.kyuubi.operation.{ArrayFetchIterator, OperationHandle, OperationState}
 import org.apache.kyuubi.operation.log.OperationLog
 import org.apache.kyuubi.session.Session
@@ -119,7 +120,8 @@ class ExecuteScala(
               }
             }
           case Error =>
-            throw KyuubiSQLException(s"Interpret error:\n$statement\n ${repl.getOutput}")
+            throw KyuubiSQLException(s"Interpret error:\n" +
+              s"${JsonUtils.toPrettyJson(Map("code" -> statement, "response" -> repl.getOutput))}")
           case Incomplete =>
             throw KyuubiSQLException(s"Incomplete code:\n$statement")
         }
